wavelynk_employee_system/requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
sqlalchemy
bcrypt
python-dotenv
reportlab
holidays
extra-streamlit-components
st-cookies-manager